+++ CHANGELOG.md
@@ -5,7 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-[Unreleased]: https://github.com/zombiezen/go-sqlite/compare/v0.10.0...main
+[Unreleased]: https://github.com/zombiezen/go-sqlite/compare/v0.10.1...main
+
+## [0.10.1][] - 2022-07-17
+
+Version 0.10.1 fixes a bug in user-defined window functions.
+Special thanks to Jan Mercl for assistance in debugging this issue.
+
+[0.10.1]: https://github.com/zombiezen/go-sqlite/releases/tag/v0.10.1
+
+### Fixed
+
+- `AggregateFinal` is now called correctly at the end of window functions' usages.
 
 ## [0.10.0][] - 2022-07-10